+++ doc/changes.rst
@@ -11,6 +11,7 @@ Version 2019.6
 * :func:`Indexer` renamed to :func:`~pystella.index_fields`.
 * :meth:`Sector.get_args` deprecated in favor of using
   :func:`~pystella.get_field_args`.
+* Add :class:`~pystella.Histogrammer` and :class:`~pystella.FieldHistogrammer`.
 * Various improvements and bug fixes.
 
 Version 2019.5